requirements.txt
absl-py==0.15.0
astroid==2.3.3
astunparse==1.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
colorama==0.4.3
cycler==0.11.0
flatbuffers==1.12
fonttools==4.28.1
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
imageio==2.10.1
isort==4.3.21
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
Markdown==3.3.4
matplotlib==3.5.0
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-scm==6.3.2
six==1.15.0
soupsieve==2.2.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
tomli==1.2.2
typing-extensions==3.7.4.3
urllib3==1.26.7
virtualenv==16.7.8
Werkzeug==2.0.2
wrapt==1.12.1
